Move all data fields on responses to a model
It didn't really make sense that a response has a concept of fields but
then might call to a model which does exactly the same thing but in a
slightly different way.

Instead, the generic response class can pass the data into a model so
there is one way to manage data. For convenience, the model fields are
still mapped to methods on the response so the consumer experience is
identical even if the implementation details are quite different