Allow models to accept symbol keys
While the API is only ever going to respond with string keys, allowing
symbols makes it easier for consumers to test their code. For example,
someone could use a tool like FactoryBot using the attributes_for method
to construct a hash of values that represent the data that would be
loaded into the model. Doing this would give you a hash with symbol keys
but you could use with_indifferent_access on the hash and the
DearInventory::Model wouldn't work as if it had received real data