Drop "Z" from date time format
Cin7 Core has introduced a bug where it cannot parse the "Z" timezone
when passing an UpdatedSince date to the saleList endpoint.

I have filed a bug with Cin7 since the API docs state that the date
should follow the ISO 8601 standard—and other parameters still accept
the "Z"—but dropping the "Z" seems to allow the endpoint to work.